fix(tracing): make OpenTelemetry an optional runtime dependency
`import scorecard_ai` previously failed with `ModuleNotFoundError: No
module named 'opentelemetry'` unless the optional `[otel]` extra was
installed, because the top-level package eagerly imported
`lib/wrap_llms.py`, which imported opentelemetry at module load time.
This made the documented `pip install scorecard-ai` broken.

Move the OpenTelemetry-backed implementation into a private
`_wrap_llms_otel` module and turn `wrap_llms` into a thin shim whose
`wrap`/`wrap_openai`/`wrap_anthropic` lazily import the implementation on
first use. When opentelemetry is missing, calling a wrapper now raises a
clear ImportError pointing to `pip install 'scorecard-ai[otel]'` instead
of breaking import of the whole SDK.
